% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{firstLetterCap}
\alias{firstLetterCap}
\title{Capitalize first letter of a string.}
\usage{
firstLetterCap(x)
}
\arguments{
\item{x}{Text string}
}
\value{
Text string
}
\description{
Capitalize first letter of a string.
}
\seealso{
Other Utility functions: \code{\link{findAbsMin}},
  \code{\link{find_n_neighbors}},
  \code{\link{getArrowPos}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{getRatioCoords}},
  \code{\link{group_sort}}, \code{\link{inch2coords}},
  \code{\link{isColor}}, \code{\link{list2str}},
  \code{\link{move_n_point}}, \code{\link{orderBoxplot}},
  \code{\link{se}}, \code{\link{sortGroups}}
}
