% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getArrowPos}
\alias{getArrowPos}
\title{Converts coordinates in current plot region to device positions (in inch).}
\usage{
getArrowPos(x, y, units = c("coords", "prop"), dev = c("plot", "figure",
  "panel"))
}
\arguments{
\item{x}{Numeric: x coordinate(s)}

\item{y}{Numeric: y coordinate(s)}

\item{units}{Coordinates (default) or proportions with respect to 
plot region.}

\item{dev}{x and y position are measured with respect to the 
plot region (default), figure panel, or device.}
}
\value{
list
}
\description{
Converts coordinates in current plot region to device positions (in inch).
}
\seealso{
Other Utility functions: \code{\link{findAbsMin}},
  \code{\link{find_n_neighbors}},
  \code{\link{firstLetterCap}}, \code{\link{getDec}},
  \code{\link{getRange}}, \code{\link{getRatioCoords}},
  \code{\link{group_sort}}, \code{\link{inch2coords}},
  \code{\link{isColor}}, \code{\link{list2str}},
  \code{\link{move_n_point}}, \code{\link{orderBoxplot}},
  \code{\link{se}}, \code{\link{sortGroups}}
}
