% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img}
\alias{img}
\title{Image object.}
\format{A list with 2 elements:
\describe{
  \item{\code{image}}{Matrix of 599 x 564 representing the image.}
  \item{\code{col}}{Vector with colors used in the image.}
}}
\usage{
img
}
\description{
Map of the Netherlands, png image by Silver Spoon on Wikipedia:
https://nl.wikipedia.org/wiki/Bestand:Blank_map_of_the_Netherlands.svg
}
\author{
Jacolien van Rij
}
\keyword{datasets}

