% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{inch2coords}
\alias{inch2coords}
\title{Convert device position (inch) to coordinates in current plot region.}
\usage{
inch2coords(xpos, ypos = NULL, simplify = FALSE)
}
\arguments{
\item{xpos}{x position in device, inches between position and left side 
of device.
 When defined as two-number vector, x- and y-position as measured from 
bottomleft corner of device.}

\item{ypos}{y position (in inches) from bottom of device.}

\item{simplify}{Logical: whether or not to output a vector instead of a list.}
}
\value{
list or 2-number vector
}
\description{
Convert device position (inch) to coordinates in current plot region.
}
\seealso{
Other Utility functions: 
\code{\link{findAbsMin}()},
\code{\link{find_n_neighbors}()},
\code{\link{firstLetterCap}()},
\code{\link{getArrowPos}()},
\code{\link{getDec}()},
\code{\link{getRange}()},
\code{\link{getRatioCoords}()},
\code{\link{get_palette}()},
\code{\link{group_sort}()},
\code{\link{isColor}()},
\code{\link{list2str}()},
\code{\link{move_n_point}()},
\code{\link{orderBoxplot}()},
\code{\link{se}()},
\code{\link{sortGroups}()}
}
\concept{Utility functions}
